Initial workaround for loginctl caching problem
There is a bug (reported to Informatics) that causes the gnome keychain
to hang around after you shut down. This results in loginctl reporting
phantom users as logged in... multiple users on seat0.

This workaround causes machines to only report users with the largest
session number. Whilst this fix will resolve the problem of incorrect
users being reported, it does not resolve the problem of users being
reported when no such user is actually logged in.